Moved tmp methods into separate module
