+++ lib/paperclip.rb
@@ -40,6 +40,7 @@ require 'paperclip/interpolations/plural_cache'
 require 'paperclip/interpolations'
 require 'paperclip/tempfile_factory'
 require 'paperclip/style'
+require 'paperclip/tmp_saveable'
 require 'paperclip/attachment'
 require 'paperclip/storage'
 require 'paperclip/callbacks'
