+++ lib/paperclip/attachment.rb
@@ -9,6 +9,8 @@ module Paperclip
   # when the model saves, deletes when the model is destroyed, and processes
   # the file upon assignment.
   class Attachment
+    include ::Paperclip::TmpSaveable
+
     def self.default_options
       @default_options ||= {
         :convert_options       => {},
@@ -151,10 +153,6 @@ module Paperclip
       end
     end
 
-    def tmp_url(style_name = default_style, options = {})
-      @url_generator.for(style_name, default_options.merge(options).merge(:tmp => true))
-    end
-
     def default_options
       {
         :timestamp => @options[:use_timestamp],
@@ -182,15 +180,6 @@ module Paperclip
         path = original_filename.nil? ? nil : interpolate(path_option, style_name)
         unescape(path)
       end
-    end
-
-    # Returns the path of the temporary upload as defined by the :url_path option. If the
-    # file is stored in the filesystem the path refers to the path of the file
-    # on disk. If the file is stored in S3, the path is the "key" part of the
-    # URL, and the :bucket option refers to the S3 bucket.
-    def tmp_path(style_name = default_style)
-      path = original_filename.nil? || tmp_id.nil? ? nil : interpolate(tmp_path_option, style_name)
-      unescape(path)
     end
 
     # :nodoc:
@@ -257,27 +246,6 @@ module Paperclip
       @dirty
     end
 
-    # Processes and saves the current file at the location determined by tmp_path.
-    # Also serializes this object in a known location so it can be used to find the saved tmp files later.
-    # If there is no current file or tmp_id, does nothing.
-    def save_tmp
-      return unless tmp_id.present? && original_filename.present?
-      FileUtils.mkdir_p(File.dirname(tmp_serialize_path))
-      File.open(tmp_serialize_path, 'w') { |f| f.write(YAML::dump(self)) }
-      save(tmp: true)
-    end
-
-    # Retrieves a saved temporary Attachment object.
-    def matching_saved_tmp
-      if tmp_id.present? && File.exists?(tmp_serialize_path)
-        YAML::load(File.read(tmp_serialize_path)).tap do |tmp|
-          tmp.send(:initialize_storage)
-        end
-      else
-        nil
-      end
-    end
-
     # Saves the file, if there are no errors. If there are, it flushes them to
     # the instance's errors and returns false, cancelling the save.
     def save(tmp: false)
@@ -290,19 +258,6 @@ module Paperclip
       flush_writes(tmp: tmp)
       @dirty = false
       true
-    end
-
-    # If a matching saved temp upload exists, assigns that as the file for this model.
-    def copy_saved_tmp_if_appropriate
-      if saved_tmp = matching_saved_tmp
-        path = saved_tmp.tmp_path(:original)
-        unless dirty?
-          if @queued_for_delete.empty? && File.exists?(path)
-            assign(File.open(path))
-          end
-          saved_tmp.delete_tmp
-        end
-      end
     end
 
     # Clears out the attachment. Has the same effect as previously assigning
@@ -449,10 +404,6 @@ module Paperclip
 
     def path_option
       @options[:path].respond_to?(:call) ? @options[:path].call(self) : @options[:path]
-    end
-
-    def tmp_path_option
-      @options[:tmp_path].respond_to?(:call) ? @options[:tmp_path].call(self) : @options[:tmp_path]
     end
 
     def active_validator_classes
@@ -679,12 +630,6 @@ module Paperclip
     # Calls `unescape` method if available
     def unescape(path)
       path.respond_to?(:unescape) ? path.unescape : path
-    end
-
-    # Gets the path where an Attachment with tmp_id matching our current tmp_id should be serialized
-    # to for later retrieval and use.
-    def tmp_serialize_path
-      "#{Rails.root}/tmp/attachments/#{tmp_id}.yml"
     end
   end
 end
